Networking and IAM updates along with cleanup of output
Build out role for use with Atlantis, added redirect for git and built out better output for useful info. 

Redirect was needed as gitlab was sending http link back to Atlantis. Redirect is happening at load balancer and pushing a 301 with url and path. 

IAM role was developed so that Atlantis would use IAM credentials which are auto generated by the project and assume a role based on account.

Output was lacking any useful information. Still need to do some minor cleanup but output now has useful information on the Bastion, Gitlab servers and Secrets of the project.